requirements/base.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.6.2
flake8==7.1.0
flake8-quotes==3.4.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.5
mypy==1.10.0
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
pyflakes==3.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
yarl==1.9.4
pyyaml==6.0.1
types-pyyaml==6.0.12.20240311
